backend/requirements.txt
fastapi[standard]==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
alembic==1.15.2
pgvector==0.4.2
pydantic==2.11.3
pydantic-settings==2.9.1
arq==0.26.1
redis==5.3.0
httpx==0.28.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.4.0
python-multipart==0.0.20
sse-starlette==2.2.1
